pyproject.toml
[project]
name = "improvedscreening"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "nicegui>=2.24.1",
    "openai>=1.107.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
